requirements.txt
certifi==2019.6.16
chardet==3.0.4
eyeD3==0.8.10
feedgen==0.7.0
idna==2.8
lxml==4.4.1
pydub==0.23.1
python-dateutil==2.8.0
python-magic==0.4.15
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
urllib3==1.25.3
